refactor(lotto): Int 에서 Long 으로 정수 범위 증가

+++ src/main/kotlin/lotto/domain/WinningResult.kt
@@ -6,8 +6,8 @@ class WinningResult(val value: HashMap<Rank, Int>) {
     fun getYieldRate(purchaseMoney: PurchaseMoney): Double =
         getTotalPrizeMoney() / purchaseMoney.value.toDouble()
 
-    private fun getTotalPrizeMoney(): Int =
-        value.keys.sumOf { rank -> rank.prizeMoney * value[rank]!! }
+    private fun getTotalPrizeMoney(): Long =
+        value.keys.sumOf { rank -> rank.prizeMoney * value[rank]!!.toLong() }
 
     companion object {
         fun from(ranks: List<Rank>): WinningResult {
